feat: support non-Git URL app names (#2353)
* feat(pkg/gomodulepath): change parse to support single name paths

* feat(pkg/gomodulepath): add function to extract user and repo names

* fix(pkg/giturl): changed to handle URL with scheme

The previous implementation was tested using an URL with scheme but CLI
is using URIs that comes from Go module names that don't have a scheme
as prefix. This implementation allows for both cases.

* refactor: improve gomodulepath pkg and tests

* feat: add support for non git module names to scaffolder service

* feat: add support for non git module names to cosmosgen pkg

* test: add integration test to scaffold a chain using a name

* test: add integration test to serve a chain scaffolded using a name

* feat(pkg/protopath): add function to format protocol buffer package names

* refactor: fix proto package names for scaffold chain & module commands

* refactor: fix proto package names for scaffolded types commands

* refactor: fix proto package names for scaffold band command

* refactor: add missing formatPackageName context definitions

* chore: remove unused formatOwnerName definition from app template

* chore: fix documentation rebranding typo

* refactor: remove redundant app name from proto file templates

This change results in proto file package names like "owner.repo"
instead of "owner.repo.repo" when a URI is used as app name during the
initial chain scaffolding. Or for example "owner.module" instead of
"owner.repo.module" when a module is created.

* refactor(pkg/gomodulepath): changed user and repo name extraction func

The function was renamed to be shorter and the extraction functionality
changed to use the last two path elements as user and repository names,
with the exception of short paths where the names are the same.

* refactor(pkg/protopath): change proto package name formatting arguments

Using the arguments is more intuitive and works better with the data
extracted from the Go module path used to scaffold the chain.

* chore: fix formatting

* refactor: change proto templates to use the updated pkg name formatter

* refactor: remove unused formatOwnerName references

* chore(pkg/gomodulepath): rename private function for clarity

* refactor: change base API URL to match new module paths

* fix: fix invalid vuex store module paths

* chore: review fixes

* chore: change integration scaffolder func to use appname as is

* test: change integration tests to explicitly use GitHub URIs

* test: change name based scaffolding integration tests to use Env.Scaffold

* chore: change existing app integration test to call Env.Scaffold

* refactor: removed the unused giturl package

* chore: fix broken cosmosanalysis test

* fix: change Env.Scaffold to only use the app name for directory paths

* refactor: rewrite gomodulepath ExtractUserRepoNames to ExtractAppPath

* refactor(pkg/cosmosgen): change pkg to use gomodulepath.ExtractAppPath

* refactor: remove FormatPackageName from protopath pkg

The function is now located in the module package within the
`ignite/templates/` path and it was renamed to ProtoPackageName.

* refactor: replace owner name references with app module paths

The owner name shouldn't be used anymore because the Go module paths
might be defined as a single name or a URI with a single path element
that only describes the app name, so in those cases there is no owner
definition.

App module paths are used instead of the owner name.

App module paths are extracted from the Go module path, e.g.:
 - github.com/username/reponame => "username/reponame"
 - domain.com/reponame => "reponame"
 - reponame => "reponame"
 - namespace/reponame => "namespace/reponame"
 - a/b/c/namespace/reponame => "namespace/reponame"

* chore: remove unused test file

* chore: fix broken Vuex store module path test

Co-authored-by: İlker G. Öztürk <ilkergoktugozturk@gmail.com>
Former-commit-id: 8631f30f176938c20ab581917ef17acb966a23d2