feat(cmd): introduce node command (#2669)
* add node command with bank query and send

Bounty address: cosmos1zra3wz596n0qc898myceka3q23x9rk8fuw65c7

* Code review cleanup

* Code review fix

* fix merge

* refactor

* Merge branch 'develop' into issue-2489-node-banks-cmds

Fixed a bunch of conflicts with recent renaming

* wip: fix integration tests due to recent refac

* test: finish apply refac in tests

* Fix integration test on node command

* various improvments in node integration test

* test: read test.v to add more trace on step execution

That helps when some commands are not running properly.

* fix: spinner hides os keyring prompt

* Ensure bank send can take account name or address

Also simplify the logic by removing the read of the --from flag, which
is no longer used in this command.

BroadcastTx now takes a cosmosaccount.Account instead of just an account
name. Since ensure the account comes from the keyring, and avoid the
BroadcastTx impl to re-check that (because the keyring is always
checked before the call to BroadcastTx).

* fix after rebase

* fix: bank balance can take an address absent of the keyring

* fix tx broadcast error message

* add fees flag to node tx bank send cmd

fees are required, at least in cosmos-hub to send funds to an other
account. W/o fees the transaction returns this message:
error code: '13' msg: 'insufficient fees; got:  required: 8uatom:
insufficient fee'

Also increase the additional gas amount added to the simulated gas,
because I got some insufficient gas error with the previous value.

* Change tx BroadcastMode to BroadcastSync

Previous mode BroadcastBlock alweus triggered a timeout error, even if
the tx was finally accepted in a block.

RPC error -32603 - Internal error: timed out waiting for tx to be
included in a block

* fix linter

* fix other lint

* improve lookupAddress error report

* fix typo

* wip

* simplify

* add assertBankBalanceOutput

* wip find a way to wait for block height

* test: replace time.Sleep with app.WaitNBlocks

* gofumpt

* fix integration test

* BroadcastBlock is deprecated

* simplify node query

* feat: move WaitForBlock methods in cosmosclient

* fix: revert usage of BroadcastSync mode

* docs: adapt blog tutorial to cosmosclient changes

* feat: add --broadcasd-mode flag to node tx command

The default value is still block but the flag can change to sync in
order to avoid timeout when the tx is broadcasted to busy nodes.

Also add a `node query tx` so the user can check when his tx is
included in a block.

* comments

* style: consolidate arg names

* fix: add port to default node

* merge const decls

* fix merge error

* fix after merge

* fix after merge

Co-authored-by: Gjermund Bjaanes <gjermund@empower.eco>
Co-authored-by: İlker G. Öztürk <ilkergoktugozturk@gmail.com>
Co-authored-by: Alex Johnson <alex@shmeeload.xyz>
Former-commit-id: 781e90a5e43aa83d3dbe9179eba107cd140711c9