fix(cosmosclient): call the faucet prior to create the tx (#2781)
Fix #2775

This change restores the behavior we used to have in 0.23, and which has
changed with the refac of cosmosclient with the `node` command feat.

Calling the faucet (when it is specifically enabled) will ensure the
account that broadcasts the tx exists in the target chain.

Because the `node` command has a `--generate-only` flag, and because we
don't want to create a new account when it's set, this flag is now
propagated to the cosmosclient struct. Therefore, the faucet isn't
called when the flag is set. This doesn't affect the `network` command
because it doesn't use that flag.

* test link issue in CL

* gh doesn't autolink issues in repo's file

* Update changelog.md

Co-authored-by: Lucas Btd <lucas.bertrand.22@gmail.com>

* Solve TODO: validate tx msgs

Co-authored-by: Lucas Btd <lucas.bertrand.22@gmail.com>
Former-commit-id: eb64d7350def0c67c3965bbf49035e05ebc8df76