test: cosmosclient (part1) (#2789)
* fix(cosmosclient): call the faucet prior to create the tx

Fix #2775

This change restores the behavior we used to have in 0.23, and which has
changed with the refac of cosmosclient with the `node` command feat.

Calling the faucet (when it is specifically enabled) will ensure the
account that broadcasts the tx exists in the target chain.

Because the `node` command has a `--generate-only` flag, and because we
don't want to create a new account when it's set, this flag is now
propagated to the cosmosclient struct. Therefore, the faucet isn't
called when the flag is set. This doesn't affect the `network` command
because it doesn't use that flag.

* test link issue in CL

* gh doesn't autolink issues in repo's file

* Update changelog.md

Co-authored-by: Lucas Btd <lucas.bertrand.22@gmail.com>

* Solve TODO: validate tx msgs

* test: cosmosclient.New

* test: cosmosclient.WaitForBlockHeight

* wip: test cosmosclient.Account

* test: cosmosclient.Account

* test: improve cosmosclient.WaitForBlockHeight

* test: cosmosclient.Address

* test: cosmosclient.Status

* refac: cosmosclient.Address

- dont need to use sdkconfig prefix since we have can just call
  account.Address(prefix)
- improve error message by using GetByName directly

* refac: no need for private account func

* restore cosmosaccount.Keyring public

* style: rename var

* add CL entry

* Update changelog.md

Co-authored-by: Alex Johnson <alex@shmeeload.xyz>

* style: var names

* merged too fast

* style: remove WithWaitBlockDuration option

In favor of a parameter in WaitForBlockHeight

* oups

* use black box testing only

Co-authored-by: Lucas Btd <lucas.bertrand.22@gmail.com>
Co-authored-by: Alex Johnson <alex@shmeeload.xyz>
Former-commit-id: 4b4ad377c33b404af9717ada5e3366afdbf87669