docs: `network` command (#2890)
* docs: `network` command

I've decided to have the most straightforward tutorial in the command description, mainly because users need to know what to look for in the documentation. It may be extracted into a separate article in the docs. In that case, we'll still need to have a link to that article in this cmd doc.

* Update ignite/cmd/network.go

Co-authored-by: Lucas Bertrand <lucas.bertrand.22@gmail.com>

* format

Co-authored-by: Lucas Bertrand <lucas.bertrand.22@gmail.com>
Co-authored-by: Alex Johnson <alex@shmeeload.xyz>
Former-commit-id: 2a3c5b6326d57be39610c04a89b4ec5fbd3761f7