feat(pkg): add support to query and save TXs and events (#2528)
* feat(pkg/cosmosclient): add method to fetch block transactions

* feat(pkg/cosmosclient): add method to collect txs from multiple blocks

* review changes

* refactor(pkg/comosclient): simplify TX and improve getter method

* feat: add cosmosmetric package

* feat(pkg/cosmosmetric): add adapter for postgres data backend

* change postgres adapter to use a DB getter

* feat: add schema support to the postgres adapter

* add more context to error messages

* chore: update "golang.org/x/sync" dependency

* feat(pkg/cosmosmetric): add raw TX save support

TXs are encoded as JSON and saved in case they need to be recreated in
the future.

Save now saves transactions in parallel instead of sequentially.

* refactor(pkg/cosmosclient): change RPC client type to an interface

The change makes it possible to write tests for calls to Tendermint RPC.

* chore(pkg/cosmosmetric): change tx collector interface name

* test(pkg/cosmosclient): add tests for Client.GetBlockTXs()

* chore(pkg/cosmosmetric): renamed SetupSchema to Init

* feat(pkg/cosmosclient): add testutil package

* test(pkg/cosmosclient): add tests for Client.CollectTXs()

* test(pkg/cosmosclient): change GetBlockTXs tests to use testutil pkg

* chore(pkg/cosmosclient): change TestCollectTXs to not use generic args

* test(pkg/cosmosclient): add test for UnmarshallEvents

* test(pkg/cosmosmetric): add mocks

* test(pkg/cosmosmetric): add collector tests

* chore(pkg/cosmosmetric): make default postgres settings public

* test(pkg/cosmosclient): change TX test to be explicitly public

* test(pkg/cosmosmetric): change test to be explicitly public

* test(pkg/cosmosclient): change client tests to be explicitly public

* test(pkg/cosmosclient): add custom tendermint RPC client mocks

* refactor: change TX to read parsed events

The refactor was required because the transaction log might contain
plaintext in some cases, for example when there are not enough funds to
pay the transaction fees.

The parsed Tendermint search result events are used instead. These
events contain JSON encoded values that are unquoted, including the
string values, so all event values should be parsed to fix this
inconsistency.

* feat(pkg/cosmosmetric): add support for queries

* chore: code change recommended by CI linter

* feat(pkg/cosmosmetric): add support for managing separate SQL schemas

This feature lets the postgres adapter keep the schema versioning
functinality while allowing external apps to also keep track of their
schemas. This is useful to let the apps create views or functions to
improve how TX and events are queried moving the loginc to the database
instead of the app.

* chore: remove lost print

* refactor(pkg/cosmosmetric): change schemas to create the database table

* refactor: changed postgres schema to simplify custom queries

The denormalized version whithout the event table needed many fields to
join the attributes and the resulting queries were less optimal that
using event IDs to gather the event attributes.

This changeset also saves transactions syncrhonously because saving
multiple TXs and events within the same transaction asynchronusly was
causing issues with the event insertion (auto ID pg lib issue) and some
times also the database connection.

* refactor: change schemas to create table and update version

This changeset allows removing the BEGIN/COMMIT and INSERT version
statements from the schema script files. The Schema will automatically
add the INSERT statement with the right version and wrap the script in a
transaction.

* refactor: improve postgres adapter to deal with schema versions

The update schema method now creates the schema table if it doesn't
exist and reads the lates version before applying the new schemas.

* refactor: change query to be more generic and better deal with calls

Query constructor changed into two, one for entity based queries and
another for call queries. Call queries are used when a view or function
must be called instead of the standard entities.

The query field definitions were moved to the adapter package so the
query module can be generic.

* refactor: change postgres filters to allow custom field filtering

This is required for call queries because the field names used in the
views or functions are not the standard TX, event and attr ones.

* refactor: change postgres parsers to work with the latest changes

The parser now joins the event and attributes because the single table
was normalized to two tables to optimize call queries.

* chore: change const definition to group by prefix

* refactor: change postgres filters to be more flexible

These changes make the Filter more generic to avoid defining extra types
to customize the filters behavior.

* refactor: rename cosmosmetric package to cosmostxcollector

* chore: change type for query page and page size

* fix: invalid Query paging states

* test(pkg/cosmostxcollector): add query tests

* test(pkg/cosmostxcollector): add postgres schemas tests

* test(pkg/cosmostxcollector): add postgres filter tests

* test(pkg/cosmostxcollector): add postgres adapter tests

* fix(pkg/cosmostxcollector): change collector to avoid zombie gophers

* fix(pkg/cosmostxcollector): change schemas to sort version before walk

* chore(pkg/cosmostxcollector): remove redundant cursor type

* fix: import paths to use `ignite` instead of `ignite-hq`

* chore: reorganize mocks

* chore: fix grouping of import statements

* chore: update changelog

* ci: correct linting

* feat: add support for querying event directly with `QueryEvents()`

* refactor: remove call queries in favor of a single query type

This change also removes the entity "fields" leaving the name of
the query fields to be chosen by the developer.

* fix: postgres query filters for slices now work

* fix: change postgres event SQL parser to check for attribute table

This way all filters for the attribute table work.

* chore: change event query to avoid attr query when there are no events

* test: add event query tests and improve other existing postgres tests

* chore: change the way `CollectTXs()` get the latest block height

Co-authored-by: Lucas Bertrand <lucas.bertrand.22@gmail.com>
Former-commit-id: faa06716ab5e0e977f14677fc5b05430b0bd1bc9